+++ test/hero_animation_test.dart
@@ -36,6 +36,98 @@ void main() {
 
     expect(find.byKey(hero1Key), findsOneWidget);
   });
+
+  testWidgets('Can use custom hero controller', (tester) async {
+    final heroController = TestHeroController();
+
+    final app = MaterialApp.router(
+      routeInformationParser: RoutemasterParser(),
+      routerDelegate: RoutemasterDelegate.builder(
+        navigatorBuilder: (context, stack) {
+          return StackNavigator(
+            stack: stack,
+            heroController: heroController,
+          );
+        },
+        routesBuilder: (context) {
+          return RouteMap(
+            routes: {
+              '/': (_) => MaterialPageOne(),
+              '/two': (_) => MaterialPageOne(),
+            },
+          );
+        },
+      ),
+    );
+    await tester.pumpWidget(app);
+    await setSystemUrl('/two');
+    await tester.pump();
+
+    expect(heroController.didPushCount, 2);
+  });
+
+  testWidgets('Can swap custom hero controller', (tester) async {
+    final heroController1 = TestHeroController();
+    final heroController2 = TestHeroController();
+
+    Widget buildApp(HeroController controller) {
+      return MaterialApp.router(
+        routeInformationParser: RoutemasterParser(),
+        routerDelegate: RoutemasterDelegate.builder(
+          navigatorBuilder: (context, stack) {
+            return StackNavigator(
+              stack: stack,
+              heroController: controller,
+            );
+          },
+          routesBuilder: (context) {
+            return RouteMap(
+              routes: {
+                '/': (_) => MaterialPageOne(),
+                '/two': (_) => MaterialPageOne(),
+                '/two/three': (_) => MaterialPageOne(),
+              },
+            );
+          },
+        ),
+      );
+    }
+
+    await tester.pumpWidget(buildApp(heroController1));
+    await setSystemUrl('/two');
+    await tester.pump();
+    await tester.pump(kTransitionDuration);
+    expect(heroController1.didPushCount, 2);
+
+    await tester.pumpWidget(buildApp(heroController2));
+    await tester.pump();
+    await setSystemUrl('/two/three');
+    await tester.pump();
+    await tester.pump(kTransitionDuration);
+    expect(heroController1.didPushCount, 2);
+    expect(heroController2.didPushCount, 1);
+  });
+}
+
+class TestHeroController extends HeroController {
+  int didPushCount = 0;
+  int didPopCount = 0;
+
+  TestHeroController({
+    CreateRectTween? createRectTween,
+  }) : super(createRectTween: createRectTween);
+
+  @override
+  void didPush(Route<dynamic> route, Route<dynamic>? previousRoute) {
+    didPushCount++;
+    super.didPush(route, previousRoute);
+  }
+
+  @override
+  void didPop(Route<dynamic> route, Route<dynamic>? previousRoute) {
+    didPopCount++;
+    super.didPop(route, previousRoute);
+  }
 }
 
 final hero1Key = Key('hero1');